arch-power: Refactor rotate instructions
This renames the mask span fields and the rotate helper
of the base class.

Reviewed-on: https://gem5-review.googlesource.com/c/public/gem5/+/40929
Maintainer: Boris Shingarov <shingarov@labware.com>
Tested-by: kokoro <noreply+kokoro@google.com>

+++ src/arch/power/insts/integer.cc
@@ -676,7 +676,7 @@ IntRotateOp::generateDisassembly(
     }
 
     // Print the shift, mask begin and mask end
-    ss << ", " << (int) sh << ", " << mb << ", " << me;
+    ss << ", " << (int) sh << ", " << (int) mb << ", " << (int) me;
 
     return ss.str();
 }
